feat(run.watch): 实现watch命令

+++ utils/cmd.go
@@ -57,6 +57,9 @@ func WaitExecRun(config *Config) {
 }
 
 func WaitKillProcess(command *exec.Cmd) {
+	if command == nil {
+		return
+	}
 	if err := command.Process.Kill(); err != nil {
 		ThrowE(err)
 	}
